requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
sentence-transformers
faiss-cpu